dp restructured
